requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.19
gunicorn==23.0.0
aiosqlite==0.19.0

